Move helper functions to utility
